# packages/core/base/src/Http/Resources/ProfileResource.php
<?php

namespace Messi\Base\Http\Resources;

/**
 * @OA\Schema(
 *     title="ProfileResource",
 *     description="ProfileResource",
 *     @OA\Xml(
 *         name="ProfileResource"
 *     )
 * )
 * @SuppressWarnings(PHPMD)
 */
class ProfileResource extends BaseResource
{
    /**
     * @OA\Property(
     *      property="id",
     *      description="id of user",
     *      example=1
     * )
     *
     * @var int
     */
    public int $id;

    /**
     * @OA\Property(
     *      property="name",
     *      description="name of user",
     *      example="name"
     * )
     *
     * @var string
     */
    public string $name;

    /**
     * @OA\Property(
     *      property="email",
     *      description="email of user",
     *      example="email@gmail.com"
     * )
     *
     * @var string
     */
    public string $email;

    /**
     * @OA\Property(
     *      property="created_at",
     *      example="2022-09-23T00:58:57.000000Z"
     * )
     *
     * @var string
     */
    public string $created_at;

    /**
     * @OA\Property(
     *      property="updated_at",
     *      example="2022-09-23T00:58:57.000000Z"
     * )
     *
     * @var string
     */
    public string $updated_at;

    /**
     * @OA\Property(
     *      property="avatar",
     *      example="123.png"
     * )
     *
     * @var string
     */
    public string $avatar;

    /**
     * @OA\Property(
     *      property="phone",
     *      example="0123456789"
     * )
     *
     * @var string
     */
    public string $phone;

    /**
     * @OA\Property(
     *      property="email_verified_at",
     *      example="true"
     * )
     *
     * @var bool
     */
    public bool $email_verified_at;

    /**
     * @param $request
     * @return array
     */
    public function toArray($request): array
    {
        return parent::toArray($request); // TODO: Change the autogenerated stub
    }
}