# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "System API",
        "description": "System API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng nhập",
                "description": "Trả về token và thời gian hết hạn",
                "operationId": "61a13c8415eed0404168aa8988fa381e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponseSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "operationId": "349b80569ed545d6952f4629261ff962",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterResponseSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "7610a570ed177369baf23fd46911868d",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh",
                "operationId": "9e8b09aa6a92b5aded1df173448f6580",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponseSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/v1/auth/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Profile",
                "operationId": "b5a500142637ace1b622d663f932504d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProfileResource": {
                "title": "ProfileResource",
                "description": "ProfileResource",
                "properties": {
                    "id": {
                        "description": "id of user",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "name of user",
                        "type": "string",
                        "example": "name"
                    },
                    "email": {
                        "description": "email of user",
                        "type": "string",
                        "example": "email@gmail.com"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2022-09-23T00:58:57.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2022-09-23T00:58:57.000000Z"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "123.png"
                    },
                    "phone": {
                        "type": "string",
                        "example": "0123456789"
                    },
                    "email_verified_at": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProfileResource"
                }
            },
            "LoginRequestSchema": {
                "title": "LoginRequest",
                "description": "LoginRequest",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "type": "string",
                        "example": "email@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of user",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LoginRequest"
                }
            },
            "LoginResponseSchema": {
                "title": "LoginResponse",
                "description": "LoginResponse",
                "properties": {
                    "access_token": {
                        "title": "access_token",
                        "description": "access_token of user",
                        "type": "string",
                        "example": "123.456.789"
                    },
                    "token_type": {
                        "title": "token_type",
                        "description": "Access token type when call api",
                        "type": "string",
                        "example": "Bearer"
                    },
                    "expires_in": {
                        "title": "expires_in",
                        "description": "expires of token",
                        "type": "integer",
                        "example": "3600"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LoginResponse"
                }
            },
            "RegisterRequestSchema": {
                "title": "RegisterRequest",
                "description": "RegisterRequest",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "type": "string",
                        "example": "email@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of user",
                        "type": "string",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "password confirmation of user",
                        "type": "string",
                        "example": "password"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "phone of user",
                        "type": "string"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of user",
                        "type": "string",
                        "example": "name"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RegisterRequest"
                }
            },
            "RegisterResponseSchema": {
                "title": "RegisterResponse",
                "description": "RegisterResponse",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of user",
                        "type": "integer",
                        "example": "1"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "type": "string",
                        "example": "email@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of user",
                        "type": "string",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "password confirmation of user",
                        "type": "string",
                        "example": "password"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "phone of user",
                        "type": "string"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of user",
                        "type": "string",
                        "example": "name"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RegisterResponse"
                }
            }
        },
...
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "name": "bearerToken",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}